% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{Compare_Barplot}
\alias{Compare_Barplot}
\title{Compare_Barplot}
\usage{
Compare_Barplot(object, Feat, compare_to)
}
\value{

}
\description{
Compare_Barplot
}
\examples{

}
\author{
Dieter Henrik Heiland
}
