% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-family.R
\name{createSlotNFCN}
\alias{createSlotNFCN}
\title{getPoints}
\usage{
createSlotNFCN(object, source, target, group.by = "seurat_clusters")
}
\arguments{
\item{coordinates}{A data.frame will rownames as identifier and coordinates names as x and y}
}
\value{

}
\description{
getPoints
}
\examples{

}
\author{
Dieter Henrik Heiland
}
