% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-family.R
\name{findDouplets}
\alias{findDouplets}
\title{findDouplets}
\usage{
findDouplets(object, subset = F, return.object = T, doublet.rate = 0.075)
}
\value{

}
\description{
findDouplets
}
\examples{

}
\author{
Dieter Henrik Heiland
}
