% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-family.R
\name{findFunctionalEnrichment}
\alias{findFunctionalEnrichment}
\title{findFunctionalEnrichment}
\usage{
findFunctionalEnrichment(marker, top = 100)
}
\value{

}
\description{
findFunctionalEnrichment
}
\examples{

}
\author{
Dieter Henrik Heiland
}
