% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getCleaned}
\alias{getCleaned}
\title{getCleaned}
\usage{
getCleaned(df, feat, q = 0.1)
}
\arguments{
\item{coordinates}{A data.frame will rownames as identifier and coordinates names as x and y}
}
\value{

}
\description{
getCleaned
}
\examples{

}
\author{
Dieter Henrik Heiland
}
