% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getDistance}
\alias{getDistance}
\title{getDistance}
\usage{
getDistance(from, to)
}
\arguments{
\item{coordinates}{A data.frame will rownames as identifier and coordinates names as x and y}
}
\value{

}
\description{
getDistance
}
\examples{

}
\author{
Dieter Henrik Heiland
}
