% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getPoints}
\alias{getPoints}
\title{getPoints}
\usage{
getPoints(object, reduction = "umap")
}
\arguments{
\item{coordinates}{A data.frame will rownames as identifier and coordinates names as x and y}
}
\value{

}
\description{
getPoints
}
\examples{

}
\author{
Dieter Henrik Heiland
}
