% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getSource}
\alias{getSource}
\title{getSource}
\usage{
getSource(object)
}
\arguments{
\item{coordinates}{A data.frame will rownames as identifier and coordinates names as x and y}
}
\value{

}
\description{
getSource
}
\examples{

}
\author{
Dieter Henrik Heiland
}
