% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getTargets}
\alias{getTargets}
\title{getTargets}
\usage{
getTargets(object)
}
\arguments{
\item{coordinates}{A data.frame will rownames as identifier and coordinates names as x and y}
}
\value{

}
\description{
getTargets
}
\examples{

}
\author{
Dieter Henrik Heiland
}
