% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-family.R
\name{inferConnectedPathways}
\alias{inferConnectedPathways}
\title{inferConnectedPathways}
\usage{
inferConnectedPathways(
  object,
  pathway,
  q = 0.1,
  group.by = "seurat_clusters",
  smooth = T,
  smooth_span = 0.02
)
}
\value{

}
\description{
inferConnectedPathways
}
\examples{

}
\author{
Dieter Henrik Heiland
}
