% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-family.R
\name{inferNFCN}
\alias{inferNFCN}
\title{inferNFCN}
\usage{
inferNFCN(
  object,
  group.by = "seurat_clusters",
  model.span = 0.5,
  q = 0.1,
  NN = 500
)
}
\value{

}
\description{
inferNFCN
}
\examples{

}
\author{
Dieter Henrik Heiland
}
