% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-family.R
\name{inferSpatialCorrection}
\alias{inferSpatialCorrection}
\title{inferSpatialCorrection}
\usage{
inferSpatialCorrection(
  object,
  inferSpatialPosition.object,
  assay = "SCT",
  estimator = 1,
  nr.genes = 20
)
}
\value{

}
\description{
inferSpatialCorrection
}
\examples{

}
\author{
Dieter Henrik Heiland
}
