% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-family.R
\name{inferSpatialPosition}
\alias{inferSpatialPosition}
\title{inferSpatialPosition}
\usage{
inferSpatialPosition(
  object,
  spata.obj,
  group.by = "seurat_clusters",
  NN = 500,
  intergrate.group.by = F,
  RL_interaction = F
)
}
\value{

}
\description{
inferSpatialPosition
}
\examples{

}
\author{
Dieter Henrik Heiland
}
