% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotFunctionalDotplot}
\alias{plotFunctionalDotplot}
\title{plotFunctionalDotplot}
\usage{
plotFunctionalDotplot(functional, top = 30)
}
\value{

}
\description{
plotFunctionalDotplot
}
\examples{

}
\author{
Dieter Henrik Heiland
}
