% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotNFCN}
\alias{plotNFCN}
\title{plotNFCN}
\usage{
plotNFCN(
  object,
  group.by = "seurat_clusters",
  pt.size = 0.2,
  pt.alpha = 0.5,
  color.top = F,
  plot.connection = T
)
}
\value{

}
\description{
plotNFCN
}
\examples{

}
\author{
Dieter Henrik Heiland
}
