% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotNFCNDotplot}
\alias{plotNFCNDotplot}
\title{plotNFCNDotplot}
\usage{
plotNFCNDotplot(object, group.by = "seurat_clusters")
}
\value{

}
\description{
plotNFCNDotplot
}
\examples{

}
\author{
Dieter Henrik Heiland
}
