% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotNFCNSpatial3D}
\alias{plotNFCNSpatial3D}
\title{plotNFCNDimRed}
\usage{
plotNFCNSpatial3D(
  object,
  ls.objects,
  reduction = "umap",
  group.by = "seurat_clusters",
  pt.size = 0.2,
  pt.alpha = 1,
  l.width = 0.1,
  l.alpha = 0.5
)
}
\value{

}
\description{
plotNFCNDimRed
}
\examples{

}
\author{
Dieter Henrik Heiland
}
