% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family.R
\name{plotNFCNDimRed}
\alias{plotNFCNDimRed}
\title{plotNFCNDimRed}
\usage{
plotNFCNDimRed(
  object,
  reduction = "umap",
  group.by = "seurat_clusters",
  pt.size = 0.2,
  pt.alpha = 0.5,
  pt.con = c(0.4, 1e-05),
  plot.connection = T,
  correction.distance = NULL
)
}
\value{

}
\description{
plotNFCNDimRed
}
\examples{

}
\author{
Dieter Henrik Heiland
}
